% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{ARE_comparison}
\alias{ARE_comparison}
\title{Compare Average Radius of Exceedance (ARE) using different models}
\usage{
ARE_comparison(
  stations,
  center = NULL,
  U1,
  U2 = NULL,
  U3 = NULL,
  u_vec = NULL,
  names = c("Truth", "XVAE", "extGAN")
)
}
\arguments{
\item{stations}{A matrix or data frame containing the coordinates of points
in the regular spatial grid. Must be a regular grid.}

\item{center}{An optional index specifying the center of the spatial domain.
If NULL, the center is determined as the point with the minimum mean distance
to all other points.}

\item{U1}{A matrix of simulated data in their uniform scalerepresenting the "Truth" model.
Rows correspond to spatial locations, and columns correspond to simulations.}

\item{U2}{An optional matrix of data representing the emulations from a model.}

\item{U3}{An optional matrix of data representing the emulations from another model.}

\item{u_vec}{A numeric vector of quantile thresholds used for calculating exceedances.
Defaults to a sequence of values between 0 and 0.998.}

\item{names}{A character vector of length three specifying labels for the models.
Defaults to c("Truth", "XVAE", "extGAN").}
}
\value{
A ggplot object showing the ARE comparison for the three models. The
plot includes lines for mean ARE and shaded areas for 95\% confidence intervals.
}
\description{
This function calculates the Average Radius of Exceedance (ARE) for different
models based on a regular grid of spatial data in their uniform scale (i.e., copulas).
}
\details{
The function calculates the ARE for the specified models at each
quantile threshold in \code{u_vec}. The ARE is computed as the average spatial
area where the model exceeds the threshold at the center of the domain.
}
\examples{
# Example usage:
data(copulas)
ARE_comparison(stations_grid, U1=U_sim_grid, U2=U_xvae_grid, U3=U_gan_grid, names =c("Truth", "XVAE", "extGAN"))
}
