% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{calc_radius}
\alias{calc_radius}
\title{Calculate Radius to Ensure Coverage by Knots}
\usage{
calc_radius(knots, stations)
}
\arguments{
\item{knots}{A matrix or dataframe of coordinates representing knot locations.}

\item{stations}{A matrix or dataframe of coordinates representing station locations.}
}
\value{
A numeric scalar representing the radius required to ensure coverage.
}
\description{
This function calculates the radius required to ensure that all station locations
are covered by at least one radial basis function centered at the knot locations.
}
\details{
The function computes the Euclidean distance between every station and all knots,
determines the nearest knot for each station, and calculates the radius as twice the maximum
of these minimum distances. This ensures that every station is within the radius of influence
of at least one knot.
}
\examples{
# Example: Calculate radius to cover all stations with knots
knots <- matrix(c(0, 0, 1, 1), ncol = 2)       # Knot coordinates
stations <- matrix(c(0.5, 0.5, 2, 2), ncol = 2) # Station coordinates
calc_radius(knots, stations)

}
