% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{circ_dist}
\alias{circ_dist}
\title{Great Circle Distance Between Two Points}
\usage{
circ_dist(x, y)
}
\arguments{
\item{x}{Numeric vector of length 2, containing the longitude and latitude of the first point (in degrees).}

\item{y}{Numeric vector of length 2, containing the longitude and latitude of the second point (in degrees).}
}
\value{
A numeric value representing the great circle distance (in kilometers) between the two points.
}
\description{
Computes the great circle distance (in kilometers) between two points on the Earth's surface
given their longitude and latitude in degrees.
}
\examples{
# Distance between Paris (48.8566, 2.3522) and New York (40.7128, -74.0060)
distance <- circ_dist(c(2.3522, 48.8566), c(-74.0060, 40.7128))
print(distance)

}
