% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{circleFun}
\alias{circleFun}
\title{Generate Points of a Circle}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{Numeric vector of length 2, specifying the \code{(x, y)} coordinates of the circle's center. Default is \code{c(0, 0)}.}

\item{diameter}{Numeric value specifying the diameter of the circle. Default is \code{1}.}

\item{npoints}{Integer specifying the number of points to generate along the circle. Default is \code{100}.}
}
\value{
A data frame with columns \code{x} and \code{y}, containing the coordinates of the points on the circle.
}
\description{
Creates a data frame containing \code{npoints} evenly spaced points around a circle
given its center and diameter.
}
\examples{
# Generate points for a circle of diameter 2 centered at (1, 1)
circle <- circleFun(center = c(1, 1), diameter = 2, npoints = 50)
plot(circle, type = "l", asp = 1)

}
