% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{data_driven_knots}
\alias{data_driven_knots}
\title{Derive Data-Driven Knots for Spatio-Temporal Models}
\usage{
data_driven_knots(
  X,
  stations,
  threshold_p,
  echo = FALSE,
  start.knot.num = NULL
)
}
\arguments{
\item{X}{Matrix. Observed data, where rows correspond to spatial sites and columns to time replicates.}

\item{stations}{Data frame or matrix. Spatial coordinates of observation sites, where rows correspond to sites.}

\item{threshold_p}{Numeric. The quantile level (e.g., 0.96) used to threshold high values in \code{X}.}

\item{echo}{Logical. If \code{TRUE}, progress messages are printed during execution.}

\item{start.knot.num}{Integer. Initial number of knots. If \code{NULL}, it defaults to approximately
\code{5 * log(nrow(stations))}.}
}
\value{
Matrix. A refined set of knot locations, where rows are coordinates.
}
\description{
This function identifies data-driven knots (spatial locations) based on high values in
the input data matrix and performs clustering to determine knot locations. Knots are
used as the basis for constructing spatial basis functions in subsequent modeling steps.
}
\details{
\enumerate{
\item High values in \code{X} exceeding the specified quantile are identified for each time replicate.
\item A clustering algorithm (\code{k-means}) is applied to the spatial locations of these high values
to identify clusters, with the number of clusters determined dynamically based on the WSS.
\item The center within each cluster is chosen as a knot.
\item Additional coarse grid knots are generated, and all knot candidates are refined to ensure
adequate spacing by eliminating close points.
}
}
\examples{
# Load example data
load("example_X.RData")
knots <- data_driven_knots(X, stations, threshold_p = 0.96)
}
