% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{double_rejection_sampler}
\alias{double_rejection_sampler}
\title{Double Rejection Sampler}
\usage{
double_rejection_sampler(theta = theta, alpha = alpha)
}
\arguments{
\item{theta}{A numeric parameter that influences the distribution.}

\item{alpha}{A numeric parameter controlling the shape of the distribution.}
}
\value{
A random sample from the distribution.
}
\description{
This function implements a rejection sampling method to generate expPS variables parameterized
by \code{alpha} and \code{theta}. Based on Devroye (2009), it involves multiple stages of sampling
from auxiliary distributions and applying acceptance/rejection criteria.
}
\examples{
double_rejection_sampler(theta = 0.5, alpha = 0.7)  # Returns a sample from the distribution
}
