% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{f_H}
\alias{f_H}
\title{f_H Function}
\usage{
f_H(x, alpha = 0.7, theta = 0.02, n = 1000, log = TRUE)
}
\arguments{
\item{x}{A numeric value for which the function is evaluated.}

\item{alpha}{A numeric value controlling the shape of the function.}

\item{theta}{A numeric value influencing the transformation.}

\item{n}{The number of points for integration when \code{alpha != 0.5}.}

\item{log}{Logical. If TRUE, the result is returned on the log scale.}
}
\value{
The computed value of the function \code{f_H}.
}
\description{
Computes a function related to the expPS distribution, parameterized by \code{alpha}, \code{theta}, and \code{x}.
The function includes two main cases based on the value of \code{alpha}. When \code{alpha == 0.5}, a specific form
of the function is calculated using \code{gamma} and \code{log}. Otherwise, numerical methods (integration) are used.
}
\examples{
f_H(1.0, alpha = 0.7, theta = 0.02)  # Returns the computed value
f_H_integrand <- function(u, x){
  A <- (sin(pi*u/2)/sin(pi*u))^2
  return(x^{-2}*A*exp(-A/x))
}
integrate(f_H_integrand, lower=0, upper=1, x=x)
}
