% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{H_density}
\alias{H_density}
\title{expPS Density Calculation}
\usage{
H_density(x, alpha = alpha, delta = delta, theta = theta)
}
\arguments{
\item{x}{Numeric vector of values at which the density is to be calculated.}

\item{alpha}{Stability parameter of the stable distribution (0 < alpha < 2).}

\item{delta}{Scale parameter for the stable distribution.}

\item{theta}{Exponential decay parameter.}
}
\value{
A numeric vector containing the computed density values at each input value of \code{x}.
}
\description{
Use functions in FMStable to calculate H density, but using numerical integration implicitly.
This function leverages the FMStable package to evaluate the stable distribution.
}
\examples{
# Example usage of H_density
x <- seq(0.1, 10, by = 0.1)
alpha <- 0.7
delta <- 1
theta <- 0.5
densities <- H_density(x, alpha = alpha, delta = delta, theta = theta)
#'
}
