% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{leaky_relu}
\alias{leaky_relu}
\title{Leaky ReLU Activation Function}
\usage{
leaky_relu(x, slope)
}
\arguments{
\item{x}{A numeric matrix or array where the activation function is applied column-wise.}

\item{slope}{A numeric value specifying the slope for negative inputs. Default values are typically small (e.g., 0.01).}
}
\value{
A numeric object of the same dimensions as \code{x}, with modified values based on the Leaky ReLU activation.
}
\description{
Implements the Leaky Rectified Linear Unit (Leaky ReLU) activation function.
This function outputs the input \code{x} if it is positive, and a small slope
times \code{x} (specified by \code{slope}) if \code{x} is negative.
}
\examples{
leaky_relu(matrix(c(-1, 2, -3, 4), nrow = 2), slope = 0.01)
# Returns a matrix with values adjusted for Leaky ReLU
}
