% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{marginal_thetavec}
\alias{marginal_thetavec}
\title{Marginal Distribution Function for Dependence Measure}
\usage{
marginal_thetavec(x, L = L, theta = theta, alpha = alpha, k_l = k_l)
}
\arguments{
\item{x}{Numeric vector of values at which the marginal distribution is evaluated.}

\item{L}{Numeric parameter related to the dependence structure.}

\item{theta}{Numeric vector of parameter values for the dependence structure.}

\item{alpha}{Stability parameter of the stable distribution (0 < alpha < 2).}

\item{k_l}{Numeric parameter related to the scaling in the dependence measure.}
}
\value{
A numeric vector containing the computed marginal distribution values for each input \code{x}.
}
\description{
Computes the marginal distribution function to calculate the dependence measure
based on Equation 10 in Bopp (2020).
}
\examples{
# Example usage of marginal_thetavec
x <- seq(0.1, 10, by = 0.1)
L <- 1
theta <- 0.5
alpha <- 0.7
k_l <- 2
result <- marginal_thetavec(x, L = L, theta = theta, alpha = alpha, k_l = k_l)

}
