% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{relu}
\alias{relu}
\title{ReLU Activation Function}
\usage{
relu(x)
}
\arguments{
\item{x}{A numeric vector, matrix, or array to apply the ReLU activation.}
}
\value{
A numeric object of the same dimensions as \code{x}, with all negative values replaced by \code{0}.
}
\description{
Implements the Rectified Linear Unit (ReLU) activation function, which is widely used in neural networks.
This function outputs the input \code{x} if it is positive; otherwise, it returns \code{0}.
}
\examples{
relu(c(-3, 0, 2))  # Returns c(0, 0, 2)
}
