% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{spatial_map}
\alias{spatial_map}
\title{Create a Spatial Map over Contiguous United States with Discrete Data Ranges}
\usage{
spatial_map(
  stations,
  var = NULL,
  pal = RColorBrewer::brewer.pal(9, "OrRd"),
  title = "spatial map",
  legend.name = "val",
  show.legend = TRUE,
  show.color = TRUE,
  show.axis.y = TRUE,
  xlab = "x",
  ylab = "y",
  brks.round = 2,
  tight.brks = FALSE,
  conus_fill = "white",
  border.wd = 0.2,
  pt.size = 3,
  shp = 16,
  range = NULL,
  q25 = NULL,
  q75 = NULL,
  raster = TRUE,
  aspect_ratio = 1
)
}
\arguments{
\item{stations}{A data frame with columns \code{x} and \code{y}, representing spatial coordinates of points to plot.}

\item{var}{Optional numeric vector to map values to colors for points or raster layers.}

\item{pal}{Color palette for the map (default is \code{RColorBrewer::brewer.pal(9,"OrRd")}).}

\item{title}{Title of the plot.}

\item{legend.name}{Name of the legend for the variable.}

\item{show.legend}{Logical; whether to display the legend (default is \code{TRUE}).}

\item{show.color}{Logical; whether to use color mapping (default is \code{TRUE} if \code{var} is provided).}

\item{show.axis.y}{Logical; whether to show the y-axis labels (default is \code{TRUE}).}

\item{xlab, }{ylab Labels for the x- and y-axes (default: \code{'x'} and \code{'y'}).}

\item{brks.round}{Number of decimal places to round breakpoints (default: 2).}

\item{tight.brks}{Logical; whether to use tighter breakpoints based on quartiles (default: \code{FALSE}).}

\item{conus_fill}{Fill color for the map background (default: \code{"white"}).}

\item{border.wd}{Width of point borders (default: \code{0.2}).}

\item{pt.size}{Size of points in the plot (default: \code{3}).}

\item{shp}{Shape of points in the plot (default: \code{16}).}

\item{range}{Range of the variable values to map (default is the range of \code{var}).}

\item{q25, }{q75 Optional quartiles for tight breakpoints.}

\item{raster}{Logical; whether to display a raster map instead of points (default: \code{TRUE}).}

\item{aspect_ratio}{Aspect ratio of the plot (default: \code{1}).}
}
\value{
A ggplot object representing the spatial map.
}
\description{
This function generates spatial maps using ggplot2, either with colored points or raster layers, for visualizing spatial data.
}
\examples{
stations <- data.frame(x = runif(100), y = runif(100), var = rnorm(100))
spatial_map(stations, var = stations$var, title = "Example Spatial Map")
}
