% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XVAE_utils.R
\name{Zolo_A}
\alias{Zolo_A}
\title{Zolo_A Function}
\usage{
Zolo_A(u, alpha = 0.7)
}
\arguments{
\item{u}{A numeric vector or value. The input for the sine functions.}

\item{alpha}{A numeric value, usually between 0 and 1, controlling the weight in the transformation.}
}
\value{
The result of the transformation applied to \code{u} and \code{alpha}.
}
\description{
Implements a mathematical transformation based on sine functions. This is used to model
specific stochastic processes and is parameterized by \code{u} and \code{alpha}.
}
\examples{
Zolo_A(1, alpha = 0.7)  # Returns the transformed value
}
